requirements.txt
wxpython
audio-separator[gpu]
requests
packaging
neuraloperator==1.0.2
safetensors
einops
# Note: torch will be installed explicitly in install.bat to target nightly/cu128
